+++ lib/Auth/JwtUtils.php
@@ -14,7 +14,7 @@ class JwtUtils
     /** @var string - using TestJewUtils */
     protected $algorithm;
 
-    public function __construct(string $public_key, string $algorithm)
+    public function __construct(string $public_key, string $algorithm = self::ALG_RS256)
     {
         $this->public_key = $public_key;
         $this->algorithm = $algorithm;
@@ -23,5 +23,25 @@ class JwtUtils
     public function decode(string $jwt): array
     {
         return (array)JWT::decode($jwt, $this->public_key, [$this->algorithm]);
+    }
+
+    public static function encode(
+        string $iss,
+        string $aud,
+        string $private_key,
+        int $exp,
+        ?string $sub = null
+    ): string {
+        $payload = [
+            'iss' => $iss,
+            'aud' => $aud,
+            'exp' => (new \DateTime())->getTimestamp() + $exp,
+        ];
+
+        if ($sub !== null) {
+            $payload['sub'] = $sub;
+        }
+
+        return JWT::encode($payload, $private_key, self::ALG_RS256);
     }
 }