+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.16.1] - 2019-11-14
+## Added
+- Add static function `encode` in `JwtUtils`
+## Changed
+- Change visibility to protected of function `getClient` in `AbstractAwsService`
+
 ## [0.16.0] - 2019-11-07
 ## Changed
 - Allow `symfony/http-foundation` ^4.3 version